refactor(date,activerecord): cite the C's own mul/div and Numeric owners; the FK-violation tests move into fixtures.test.ts
`subsecDigits` names `date_strftime.c`'s two operations and why they cannot be
written that way in JS. `Rational#div` / `#%` / `#zero?` are Numeric's, not
Rational's (`ruby -e 'Rational(1,3).method(:div).owner'`), so they cite
`numeric.c` and the `f_idiv` / `f_mod` / `f_zero_p` that dispatch to them.

`FixturesWithForeignKeyViolationsTest` moves into `fixtures.test.ts`, the file
that answers `fixtures_test.rb`, and takes the canonical
`fkObjectToPointToFixtureData` rather than an inline row.
